move cruisedesignFileContext to seperate file
